+++ src/types.ts
@@ -16,6 +16,7 @@ import { SsProfileWithCustomText } from './tools/auth0/handlers/selfServiceProfi
 import { PhoneProvider } from './tools/auth0/handlers/phoneProvider';
 import { PhoneTemplate } from './tools/auth0/handlers/phoneTemplates';
 import { NetworkACL } from './tools/auth0/handlers/networkACLs';
+import { NetworkAclKey } from './tools/auth0/handlers/networkACLKeys';
 import { UserAttributeProfile } from './tools/auth0/handlers/userAttributeProfiles';
 import { AttackProtection } from './tools/auth0/handlers/attackProtection';
 import { TokenExchangeProfile } from './tools/auth0/handlers/tokenExchangeProfiles';
@@ -163,6 +164,7 @@ export type Assets = Partial<{
   flows: Flow[] | null;
   flowVaultConnections: FlowVaultConnection[] | null;
   selfServiceProfiles: SsProfileWithCustomText[] | null;
+  networkACLKeys: NetworkAclKey[] | null;
   networkACLs: NetworkACL[] | null;
   userAttributeProfiles: UserAttributeProfile[] | null;
   userAttributeProfilesWithId: UserAttributeProfile[] | null;
@@ -231,6 +233,7 @@ export type AssetTypes =
   | 'flows'
   | 'flowVaultConnections'
   | 'selfServiceProfiles'
+  | 'networkACLKeys'
   | 'networkACLs'
   | 'userAttributeProfiles'
   | 'connectionProfiles'